+++ scripts/mainline_integration_review.rb
@@ -83,6 +83,7 @@ LANES = [
       %r{\Ainternal/runtime/appidentity/runtime_route_convergence},
       %r{\Ainternal/runtime/appidentity/runtime_service_activation_preflight},
       %r{\Ainternal/runtime/appidentity/runtime_service_binding_},
+      %r{\Ainternal/runtime/appidentity/runtime_write_gate\.go\z},
       %r{\Ainternal/runtime/appidentity/runtime_write_gate_},
       %r{\Ascripts/runtime_contract_drift_report\.rb\z},
       %r{\Ascripts/runtime_owner_candidate_smoke\.rb\z},